requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bleach==5.0.1
boto==2.49.0
boto3==1.26.165
botocore==1.29.165
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.6
django-allauth==0.51.0
django-bootstrap-modal-forms==2.2.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-dbbackup==1.80.1
django-environ==0.9.0
django-filter==22.1
django-pgtrigger==4.6.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
dnspython==2.2.1
drf-api-logger==1.1.11
drf-spectacular==0.24.2
dropbox==11.36.0
eventlet==0.33.1
exceptiongroup==1.0.0rc9
factory-boy==3.2.1
Faker==15.1.1
filelock==3.9.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.12.0
google-cloud-storage==2.5.0
google-cloud-tasks==2.12.1
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.14.0
httpie==3.2.1
identify==2.5.12
idna==3.4
importlib-resources==5.10.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
kombu==5.2.4
lambda-pkg-resources==0.0.6
MarkupSafe==2.1.1
multidict==6.0.2
nodeenv==1.7.0
oauthlib==3.2.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
postgres==4.0
pre-commit==2.21.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.8.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyproject-toml==0.0.10
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
redis==4.5.5
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
rich==12.6.0
rsa==4.9
s3transfer==0.6.1
sendgrid==6.9.7
sentry-sdk==1.15.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
stone==3.3.1
tea==0.1.7
toml==0.10.2
tomli==2.0.1
tremendous-python==2.2.3
twilio==7.16.0
types-cryptography==3.3.23.1
types-pkg-resources==0.1.3
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==3.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.20.0
vine==5.0.0
virtualenv==20.17.1
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.9.0
